refactor: message views
