+++ src/components/Chat/ChatContent/Message/View/Button/DownButton.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import DownChevronArrow from '@icon/DownChevronArrow';
+
+import BaseButton from './BaseButton';
+
+const DownButton = ({
+  onClick,
+}: {
+  onClick: React.MouseEventHandler<HTMLButtonElement>;
+}) => {
+  return <BaseButton icon={<DownChevronArrow />} onClick={onClick} />;
+};
+
+export default DownButton;
