+++ src/components/Chat/ChatContent/Message/View/Button/RefreshButton.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import RefreshIcon from '@icon/RefreshIcon';
+
+import BaseButton from './BaseButton';
+
+const RefreshButton = ({
+  onClick,
+}: {
+  onClick: React.MouseEventHandler<HTMLButtonElement>;
+}) => {
+  return <BaseButton icon={<RefreshIcon />} onClick={onClick} />;
+};
+
+export default RefreshButton;
