+++ src/components/Chat/ChatContent/Message/View/ContentView.tsx
@@ -1,0 +1,193 @@
+import React, {
+  DetailedHTMLProps,
+  HTMLAttributes,
+  memo,
+  useState,
+} from 'react';
+
+import ReactMarkdown from 'react-markdown';
+import { CodeProps, ReactMarkdownProps } from 'react-markdown/lib/ast-to-react';
+
+import rehypeKatex from 'rehype-katex';
+import rehypeHighlight from 'rehype-highlight';
+import remarkMath from 'remark-math';
+import remarkGfm from 'remark-gfm';
+import useStore from '@store/store';
+
+import TickIcon from '@icon/TickIcon';
+import CrossIcon from '@icon/CrossIcon';
+
+import useSubmit from '@hooks/useSubmit';
+
+import { ChatInterface } from '@type/chat';
+
+import { codeLanguageSubset } from '@constants/chat';
+
+import RefreshButton from './Button/RefreshButton';
+import UpButton from './Button/UpButton';
+import DownButton from './Button/DownButton';
+import CopyButton from './Button/CopyButton';
+import EditButton from './Button/EditButton';
+import DeleteButton from './Button/DeleteButton';
+import CodeBlock from '../CodeBlock';
+
+const ContentView = memo(
+  ({
+    role,
+    content,
+    setIsEdit,
+    messageIndex,
+  }: {
+    role: string;
+    content: string;
+    setIsEdit: React.Dispatch<React.SetStateAction<boolean>>;
+    messageIndex: number;
+  }) => {
+    const { handleSubmit } = useSubmit();
+    const [isDelete, setIsDelete] = useState<boolean>(false);
+    const currentChatIndex = useStore((state) => state.currentChatIndex);
+    const setChats = useStore((state) => state.setChats);
+    const lastMessageIndex = useStore((state) =>
+      state.chats ? state.chats[state.currentChatIndex].messages.length - 1 : 0
+    );
+    const inlineLatex = useStore((state) => state.inlineLatex);
+
+    const handleDelete = () => {
+      const updatedChats: ChatInterface[] = JSON.parse(
+        JSON.stringify(useStore.getState().chats)
+      );
+      updatedChats[currentChatIndex].messages.splice(messageIndex, 1);
+      setChats(updatedChats);
+    };
+
+    const handleMove = (direction: 'up' | 'down') => {
+      const updatedChats: ChatInterface[] = JSON.parse(
+        JSON.stringify(useStore.getState().chats)
+      );
+      const updatedMessages = updatedChats[currentChatIndex].messages;
+      const temp = updatedMessages[messageIndex];
+      if (direction === 'up') {
+        updatedMessages[messageIndex] = updatedMessages[messageIndex - 1];
+        updatedMessages[messageIndex - 1] = temp;
+      } else {
+        updatedMessages[messageIndex] = updatedMessages[messageIndex + 1];
+        updatedMessages[messageIndex + 1] = temp;
+      }
+      setChats(updatedChats);
+    };
+
+    const handleMoveUp = () => {
+      handleMove('up');
+    };
+
+    const handleMoveDown = () => {
+      handleMove('down');
+    };
+
+    const handleRefresh = () => {
+      const updatedChats: ChatInterface[] = JSON.parse(
+        JSON.stringify(useStore.getState().chats)
+      );
+      const updatedMessages = updatedChats[currentChatIndex].messages;
+      updatedMessages.splice(updatedMessages.length - 1, 1);
+      setChats(updatedChats);
+      handleSubmit();
+    };
+
+    const handleCopy = () => {
+      navigator.clipboard.writeText(content);
+    };
+
+    return (
+      <>
+        <div className='markdown prose w-full md:max-w-full break-words dark:prose-invert dark share-gpt-message'>
+          <ReactMarkdown
+            remarkPlugins={[
+              remarkGfm,
+              [remarkMath, { singleDollarTextMath: inlineLatex }],
+            ]}
+            rehypePlugins={[
+              rehypeKatex,
+              [
+                rehypeHighlight,
+                {
+                  detect: true,
+                  ignoreMissing: true,
+                  subset: codeLanguageSubset,
+                },
+              ],
+            ]}
+            linkTarget='_new'
+            components={{
+              code,
+              p,
+            }}
+          >
+            {content}
+          </ReactMarkdown>
+        </div>
+        <div className='flex justify-end gap-2 w-full mt-2'>
+          {isDelete || (
+            <>
+              {!useStore.getState().generating &&
+                role === 'assistant' &&
+                messageIndex === lastMessageIndex && (
+                  <RefreshButton onClick={handleRefresh} />
+                )}
+              {messageIndex !== 0 && <UpButton onClick={handleMoveUp} />}
+              {messageIndex !== lastMessageIndex && (
+                <DownButton onClick={handleMoveDown} />
+              )}
+
+              <CopyButton onClick={handleCopy} />
+              <EditButton setIsEdit={setIsEdit} />
+              <DeleteButton setIsDelete={setIsDelete} />
+            </>
+          )}
+          {isDelete && (
+            <>
+              <button
+                className='p-1 hover:text-white'
+                onClick={() => setIsDelete(false)}
+              >
+                <CrossIcon />
+              </button>
+              <button className='p-1 hover:text-white' onClick={handleDelete}>
+                <TickIcon />
+              </button>
+            </>
+          )}
+        </div>
+      </>
+    );
+  }
+);
+
+const code = memo((props: CodeProps) => {
+  const { inline, className, children } = props;
+  const match = /language-(\w+)/.exec(className || '');
+  const lang = match && match[1];
+
+  if (inline) {
+    return <code className={className}>{children}</code>;
+  } else {
+    return <CodeBlock lang={lang || 'text'} codeChildren={children} />;
+  }
+});
+
+const p = memo(
+  (
+    props?: Omit<
+      DetailedHTMLProps<
+        HTMLAttributes<HTMLParagraphElement>,
+        HTMLParagraphElement
+      >,
+      'ref'
+    > &
+      ReactMarkdownProps
+  ) => {
+    return <p className='whitespace-pre-wrap'>{props?.children}</p>;
+  }
+);
+
+export default ContentView;
