+++ src/components/Chat/ChatContent/Message/View/EditView.tsx
@@ -1,0 +1,244 @@
+import React, { memo, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import useStore from '@store/store';
+
+import useSubmit from '@hooks/useSubmit';
+
+import { ChatInterface } from '@type/chat';
+
+import PopupModal from '@components/PopupModal';
+import TokenCount from '@components/TokenCount';
+import CommandPrompt from '../CommandPrompt';
+
+const EditView = ({
+  content,
+  setIsEdit,
+  messageIndex,
+  sticky,
+}: {
+  content: string;
+  setIsEdit: React.Dispatch<React.SetStateAction<boolean>>;
+  messageIndex: number;
+  sticky?: boolean;
+}) => {
+  const inputRole = useStore((state) => state.inputRole);
+  const setChats = useStore((state) => state.setChats);
+  const currentChatIndex = useStore((state) => state.currentChatIndex);
+
+  const [_content, _setContent] = useState<string>(content);
+  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
+  const textareaRef = React.createRef<HTMLTextAreaElement>();
+
+  const { t } = useTranslation();
+
+  const resetTextAreaHeight = () => {
+    if (textareaRef.current) textareaRef.current.style.height = 'auto';
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    const isMobile =
+      /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini|playbook|silk/i.test(
+        navigator.userAgent
+      );
+
+    if (e.key === 'Enter' && !isMobile && !e.nativeEvent.isComposing) {
+      const enterToSubmit = useStore.getState().enterToSubmit;
+      if (sticky) {
+        if (
+          (enterToSubmit && !e.shiftKey) ||
+          (!enterToSubmit && (e.ctrlKey || e.shiftKey))
+        ) {
+          e.preventDefault();
+          handleSaveAndSubmit();
+          resetTextAreaHeight();
+        }
+      } else {
+        if (e.ctrlKey && e.shiftKey) {
+          e.preventDefault();
+          handleSaveAndSubmit();
+          resetTextAreaHeight();
+        } else if (e.ctrlKey || e.shiftKey) handleSave();
+      }
+    }
+  };
+
+  const handleSave = () => {
+    if (sticky && (_content === '' || useStore.getState().generating)) return;
+    const updatedChats: ChatInterface[] = JSON.parse(
+      JSON.stringify(useStore.getState().chats)
+    );
+    const updatedMessages = updatedChats[currentChatIndex].messages;
+    if (sticky) {
+      updatedMessages.push({ role: inputRole, content: _content });
+      _setContent('');
+      resetTextAreaHeight();
+    } else {
+      updatedMessages[messageIndex].content = _content;
+      setIsEdit(false);
+    }
+    setChats(updatedChats);
+  };
+
+  const { handleSubmit } = useSubmit();
+  const handleSaveAndSubmit = () => {
+    if (useStore.getState().generating) return;
+    const updatedChats: ChatInterface[] = JSON.parse(
+      JSON.stringify(useStore.getState().chats)
+    );
+    const updatedMessages = updatedChats[currentChatIndex].messages;
+    if (sticky) {
+      if (_content !== '') {
+        updatedMessages.push({ role: inputRole, content: _content });
+      }
+      _setContent('');
+      resetTextAreaHeight();
+    } else {
+      updatedMessages[messageIndex].content = _content;
+      updatedChats[currentChatIndex].messages = updatedMessages.slice(
+        0,
+        messageIndex + 1
+      );
+      setIsEdit(false);
+    }
+    setChats(updatedChats);
+    handleSubmit();
+  };
+
+  useEffect(() => {
+    if (textareaRef.current) {
+      textareaRef.current.style.height = 'auto';
+      textareaRef.current.style.height = `${textareaRef.current.scrollHeight}px`;
+    }
+  }, [_content]);
+
+  useEffect(() => {
+    if (textareaRef.current) {
+      textareaRef.current.style.height = 'auto';
+      textareaRef.current.style.height = `${textareaRef.current.scrollHeight}px`;
+    }
+  }, []);
+
+  return (
+    <>
+      <div
+        className={`w-full ${
+          sticky
+            ? 'py-2 md:py-3 px-2 md:px-4 border border-black/10 bg-white dark:border-gray-900/50 dark:text-white dark:bg-gray-700 rounded-md shadow-[0_0_10px_rgba(0,0,0,0.10)] dark:shadow-[0_0_15px_rgba(0,0,0,0.10)]'
+            : ''
+        }`}
+      >
+        <textarea
+          ref={textareaRef}
+          className='m-0 resize-none rounded-lg bg-transparent overflow-y-hidden focus:ring-0 focus-visible:ring-0 leading-7 w-full placeholder:text-gray-500/40'
+          onChange={(e) => {
+            _setContent(e.target.value);
+          }}
+          value={_content}
+          placeholder={t('submitPlaceholder') as string}
+          onKeyDown={handleKeyDown}
+          rows={1}
+        ></textarea>
+      </div>
+      <EditViewButtons
+        sticky={sticky}
+        handleSaveAndSubmit={handleSaveAndSubmit}
+        handleSave={handleSave}
+        setIsModalOpen={setIsModalOpen}
+        setIsEdit={setIsEdit}
+        _setContent={_setContent}
+      />
+      {isModalOpen && (
+        <PopupModal
+          setIsModalOpen={setIsModalOpen}
+          title={t('warning') as string}
+          message={t('clearMessageWarning') as string}
+          handleConfirm={handleSaveAndSubmit}
+        />
+      )}
+    </>
+  );
+};
+
+const EditViewButtons = memo(
+  ({
+    sticky = false,
+    handleSaveAndSubmit,
+    handleSave,
+    setIsModalOpen,
+    setIsEdit,
+    _setContent,
+  }: {
+    sticky?: boolean;
+    handleSaveAndSubmit: () => void;
+    handleSave: () => void;
+    setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
+    setIsEdit: React.Dispatch<React.SetStateAction<boolean>>;
+    _setContent: React.Dispatch<React.SetStateAction<string>>;
+  }) => {
+    const { t } = useTranslation();
+    const generating = useStore.getState().generating;
+    const advancedMode = useStore((state) => state.advancedMode);
+
+    return (
+      <div className='flex'>
+        <div className='flex-1 text-center mt-2 flex justify-center'>
+          {sticky && (
+            <button
+              className={`btn relative mr-2 btn-primary ${
+                generating ? 'cursor-not-allowed opacity-40' : ''
+              }`}
+              onClick={handleSaveAndSubmit}
+            >
+              <div className='flex items-center justify-center gap-2'>
+                {t('saveAndSubmit')}
+              </div>
+            </button>
+          )}
+
+          <button
+            className={`btn relative mr-2 ${
+              sticky
+                ? `btn-neutral ${
+                    generating ? 'cursor-not-allowed opacity-40' : ''
+                  }`
+                : 'btn-primary'
+            }`}
+            onClick={handleSave}
+          >
+            <div className='flex items-center justify-center gap-2'>
+              {t('save')}
+            </div>
+          </button>
+
+          {sticky || (
+            <button
+              className='btn relative mr-2 btn-neutral'
+              onClick={() => {
+                !generating && setIsModalOpen(true);
+              }}
+            >
+              <div className='flex items-center justify-center gap-2'>
+                {t('saveAndSubmit')}
+              </div>
+            </button>
+          )}
+
+          {sticky || (
+            <button
+              className='btn relative btn-neutral'
+              onClick={() => setIsEdit(false)}
+            >
+              <div className='flex items-center justify-center gap-2'>
+                {t('cancel')}
+              </div>
+            </button>
+          )}
+        </div>
+        {sticky && advancedMode && <TokenCount />}
+        <CommandPrompt _setContent={_setContent} />
+      </div>
+    );
+  }
+);
+
+export default EditView;
